@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 """
-(c) 2012-2021 Martin Wendt; see https://github.com/mar10/pyftpsync
+(c) 2012-2022 Martin Wendt; see https://github.com/mar10/pyftpsync
 Licensed under the MIT license: https://www.opensource.org/licenses/mit-license.php
 """
 
@@ -106,7 +106,8 @@ class _Target:
         self.readonly = False
         self.dry_run = False
         self.host = None
-        self.synchronizer = None  # Set by BaseSynchronizer.__init__()
+        #: Set by BaseSynchronizer.__init__(). May be None for tree command, etc.
+        self.synchronizer = None
         self.peer = None
         self.cur_dir = None
         self.connected = False
@@ -144,7 +145,16 @@ class _Target:
         return "{}".format(self.root_dir)
 
     def is_local(self):
+        if not self.synchronizer:
+            raise RuntimeError(
+                "Unbound target: Consider calling target.is_remote(or_unbound=True)"
+            )
         return self.synchronizer.local is self
+
+    def is_remote(self, or_unbound=False):
+        if or_unbound and not self.synchronizer:
+            return True
+        return not self.is_local()
 
     def is_unbound(self):
         return self.synchronizer is None
@@ -419,7 +429,7 @@ class FsTarget(_Target):
         # root_dir = self.to_unicode(root_dir)
         root_dir = os.path.expanduser(root_dir)
         root_dir = os.path.abspath(root_dir)
-        super(FsTarget, self).__init__(root_dir, extra_opts)
+        super().__init__(root_dir, extra_opts)
         if not os.path.isdir(root_dir):
             raise ValueError("{} is not a directory.".format(root_dir))
         self.support_set_time = True
@@ -430,11 +440,11 @@ class FsTarget(_Target):
         )
 
     def open(self):
-        super(FsTarget, self).open()
+        super().open()
         self.cur_dir = self.root_dir
 
     def close(self):
-        super(FsTarget, self).close()
+        super().close()
 
     def cwd(self, dir_name):
         path = normpath_url(join_url(self.cur_dir, dir_name))
