@@ -23,12 +23,12 @@ class CliTest(_SyncTestBase):
 
     def setUp(self):
         # Call self._prepare_initial_synced_fixture():
-        super(CliTest, self).setUp()
+        super().setUp()
         self.local = get_local_test_url()
         self.remote = get_remote_test_url()
 
     def tearDown(self):
-        super(CliTest, self).tearDown()
+        super().tearDown()
 
     def test_basic(self):
         out = run_script("--version")
@@ -47,15 +47,21 @@ class CliTest(_SyncTestBase):
 
     def test_sync(self):
         out = run_script("sync", self.local, self.remote, "--dry-run")
-        assert "(DRY-RUN) Wrote 0/16 files in 7 directories, skipped: 0." in out
+        assert (
+            "(DRY-RUN) Wrote 0/16 files in 7 directories, skipped: 0, errors: 0." in out
+        )
 
     def test_upload(self):
         out = run_script("upload", self.local, self.remote, "--dry-run")
-        assert "(DRY-RUN) Wrote 0/16 files in 7 directories, skipped: 0." in out
+        assert (
+            "(DRY-RUN) Wrote 0/16 files in 7 directories, skipped: 0, errors: 0." in out
+        )
 
     def test_download(self):
         out = run_script("download", self.local, self.remote, "--dry-run")
-        assert "(DRY-RUN) Wrote 0/16 files in 7 directories, skipped: 0." in out
+        assert (
+            "(DRY-RUN) Wrote 0/16 files in 7 directories, skipped: 0, errors: 0." in out
+        )
 
 
 # ===============================================================================
