@@ -14,4 +14,4 @@ Examples
     Developmental release (to mark 3.0.0 as 'used'. Don't publish this):
         '3.0.0-dev1'
 """
-__version__ = "4.0.0-a4"
+__version__ = "4.0.1-a1"
