@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 """
-(c) 2012-2021 Martin Wendt; see https://github.com/mar10/pyftpsync
+(c) 2012-2022 Martin Wendt; see https://github.com/mar10/pyftpsync
 Licensed under the MIT license: https://www.opensource.org/licenses/mit-license.php
 """
 
@@ -24,6 +24,7 @@ def add_tree_parser(subparsers):
         "tree",
         parents=[verbose_parser, common_parser, matcher_parser, creds_parser],
         help="list target folder structure",
+        allow_abbrev=False,
     )
     parser.add_argument(
         "target",
@@ -33,7 +34,9 @@ def add_tree_parser(subparsers):
     )
     parser.add_argument("--files", action="store_true", help="list files")
     parser.add_argument(
-        "--sort", action="store_true", help="sort by name (folders before files)"
+        "--sort",
+        action="store_true",
+        help="sort by name (case insensitive, files before folders)",
     )
 
     parser.set_defaults(command=tree_handler)
@@ -42,7 +45,7 @@ def add_tree_parser(subparsers):
 
 
 def tree_handler(parser, args):
-    """Implement `scan` sub-command."""
+    """Implement `tree` sub-command."""
     opts = namespace_to_dict(args)
     opts.update({"ftp_debug": args.verbose >= 6})
     target = make_target(args.target, opts)
