@@ -5,12 +5,18 @@ Tests for pyftpsync
 import os
 import sys
 
+from build.lib.ftpsync.util import get_debug_option
+
 sys.path.append(os.path.dirname(__file__) + "/..")
 
 
 # ===============================================================================
 # Main
 # ===============================================================================
+
+PYFTPSYNC_REMOTE_URL = get_debug_option(
+    "PYFTPSYNC_DEBUG_FTP_URL", "debug", "ftp_url", "ftp://example.com/test_pyftpsync"
+)
 
 
 def do_sync():
@@ -19,7 +25,7 @@ def do_sync():
 
     local_target = make_target("~/test_pyftpsync_dl")
     # local_target = make_target("c:/tmp/test_pyftpsync")
-    remote_target = make_target("ftp://wwwendt.de/test_pyftpsync", {"ftp_debug": False})
+    remote_target = make_target(PYFTPSYNC_REMOTE_URL, {"ftp_debug": False})
     opts = {"verbose": 6, "dry_run": False, "resolve": "local"}
     s = BiDirSynchronizer(local_target, remote_target, opts)
     s.run()
@@ -31,7 +37,7 @@ def do_scan():
     class args:  # noqa: N801
         """"""
 
-    args.target = "ftp://wwwendt.de/test_pyftpsync"
+    args.target = PYFTPSYNC_REMOTE_URL
     args.list = True
     args.verbose = 4
     args.recursive = False
